--- jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest',
-  }
-}
