--- ossjs.release.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  script: 'pnpm publish --no-git-checks --access public'
-}
