+++ package.json
@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "name": "@open-draft/until",
   "version": "2.1.0",
   "description": "Gracefully handle a Promise using async/await.",
@@ -17,20 +18,18 @@
   "author": "Artem Zakharchenko <kettanaito@gmail.com>",
   "license": "MIT",
   "files": [
-    "README.md",
-    "lib"
+    "lib",
+    "src"
   ],
   "devDependencies": {
     "@ossjs/release": "^0.8.1",
-    "@types/jest": "^26.0.22",
-    "jest": "^26.6.3",
-    "ts-jest": "^26.5.4",
     "tsup": "^8.5.0",
-    "typescript": "^5.9.2"
+    "typescript": "^5.9.2",
+    "vitest": "^3.2.4"
   },
   "scripts": {
     "start": "tsc -w",
-    "test": "jest",
+    "test": "vitest",
     "build": "tsup",
     "release": "release publish"
   }
