+++ src/until.test.ts
@@ -1,60 +1,27 @@
-import { AsyncTuple, until } from './until'
+import { until } from './index.js'
 
-describe('until', () => {
-  describe('given a callback function that returns a value', () => {
-    let result: AsyncTuple
-
-    beforeAll(async () => {
-      result = await until(() => Promise.resolve('value'))
-    })
-
-    it('should return explicit null as the error', () => {
-      expect(result.error).toBe(null)
-    })
-
-    it('should return data', () => {
-      expect(result.data).toEqual('value')
-    })
+it('resolves with the value returned from the callback', async () => {
+  await expect(until(async () => 'value')).resolves.toEqual({
+    error: null,
+    data: 'value',
   })
 
-  describe('given a callback function that throws an exception', () => {
-    const customError = new Error('Error message')
+  await expect(until(() => Promise.resolve('value'))).resolves.toEqual({
+    error: null,
+    data: 'value',
+  })
+})
 
-    const run = () => {
-      return until(() => {
-        throw customError
-      })
-    }
-
-    it('should not throw when called', () => {
-      expect(run).not.toThrow()
-    })
-
-    it('should return caught error as error', async () => {
-      const { error } = await run()
-      expect(error).toEqual(customError)
-    })
-
-    it('should return explicit null as data', async () => {
-      const { data } = await run()
-      expect(data).toBe(null)
-    })
+it('resolves with the error thrown in the callback', async () => {
+  await expect(
+    until(() => Promise.reject(new Error('error'))),
+  ).resolves.toEqual({
+    error: new Error('error'),
+    data: null,
   })
 
-  describe('given a Promise that rejects', () => {
-    let result: AsyncTuple
-
-    beforeAll(async () => {
-      result = await until(() => Promise.reject(new Error('Error message')))
-    })
-
-    it('should return the rejected error', () => {
-      expect(result.error).toBeInstanceOf(Error)
-      expect(result.error).toHaveProperty('message', 'Error message')
-    })
-
-    it('should return explicit null as data', () => {
-      expect(result.data).toBeNull()
-    })
+  await expect(until(() => Promise.reject('custom reason'))).resolves.toEqual({
+    error: 'custom reason',
+    data: null,
   })
 })
